+++ eval/context_test.go
@@ -7,9 +7,13 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"reflect"
+	"strings"
 	"testing"
 
+	"github.com/google/go-cmp/cmp"
+	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/hclsimple"
+	"github.com/hashicorp/hcl/v2/hclsyntax"
 	"github.com/zclconf/go-cty/cty"
 
 	"github.com/avenga/couper/config/configload"
@@ -85,10 +89,18 @@ func TestNewHTTPContext(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(subT *testing.T) {
+			file, diags := hclsyntax.ParseConfig([]byte(tt.hcl), "test.hcl", hcl.InitialPos)
+			if diags.HasErrors() {
+				subT.Fatal(diags)
+			}
+			bufferOption := eval.MustBuffer(file.Body)
+
 			helper := test.New(subT)
 
 			req := httptest.NewRequest(tt.reqMethod, "https://couper.io/"+tt.query, tt.body)
-			*req = *req.Clone(context.WithValue(req.Context(), request.Endpoint, "couper-proxy"))
+			ctx := context.WithValue(req.Context(), request.Endpoint, "couper-proxy")
+			ctx = context.WithValue(ctx, request.BufferOptions, bufferOption)
+			*req = *req.Clone(ctx)
 
 			for k, v := range tt.reqHeader {
 				req.Header[k] = v
@@ -97,13 +109,13 @@ func TestNewHTTPContext(t *testing.T) {
 			bereq := req.Clone(context.Background())
 			beresp := newBeresp(bereq)
 
-			helper.Must(eval.SetGetBody(req, eval.BufferRequest, 512))
+			helper.Must(eval.SetGetBody(req, bufferOption, 512))
 
-			ctx := baseCtx.WithClientRequest(req).WithBeresp(beresp, cty.NilVal, false).HCLContext()
-			ctx.Functions = nil // we are not interested in a functions test
+			hclCtx := baseCtx.WithClientRequest(req).WithBeresp(beresp, cty.NilVal, false).HCLContext()
+			hclCtx.Functions = nil // we are not interested in a functions test
 
 			var resultMap map[string]cty.Value
-			_ = hclsimple.Decode(tt.name+".hcl", []byte(tt.hcl), ctx, &resultMap)
+			_ = hclsimple.Decode(tt.name+".hcl", []byte(tt.hcl), hclCtx, &resultMap)
 
 			for k, v := range tt.want {
 				cv, ok := resultMap[k]
@@ -115,6 +127,183 @@ func TestNewHTTPContext(t *testing.T) {
 				if !reflect.DeepEqual(v, result) {
 					subT.Errorf("Expected %q, got: %#v, type: %#v", v, result, cv)
 				}
+			}
+		})
+	}
+}
+
+func TestContext_buffer_parseJSON(t *testing.T) {
+	baseCtx := eval.NewDefaultContext()
+	tests := []struct {
+		name                  string
+		bufferOptions         eval.BufferOption
+		contentType           string
+		expBereqsDefBody      interface{}
+		expBereqsDefJsonBody  interface{}
+		expBereqBody          interface{}
+		expBereqJsonBody      interface{}
+		expBerespsDefBody     interface{}
+		expBerespsDefJsonBody interface{}
+		expBerespBody         interface{}
+		expBerespJsonBody     interface{}
+	}{
+		{
+			"buffer both, json-parse both, application/json",
+			eval.BufferRequest | eval.JSONParseRequest | eval.BufferResponse | eval.JSONParseResponse,
+			"application/json",
+			`{"a":"1"}`,
+			map[string]interface{}{"a": "1"},
+			`{"a":"1"}`,
+			map[string]interface{}{"a": "1"},
+			`{"b":"2"}`,
+			map[string]interface{}{"b": "2"},
+			`{"b":"2"}`,
+			map[string]interface{}{"b": "2"},
+		},
+		{
+			"buffer both, json-parse both, application/foo+json",
+			eval.BufferRequest | eval.JSONParseRequest | eval.BufferResponse | eval.JSONParseResponse,
+			"application/foo+json",
+			`{"a":"1"}`,
+			map[string]interface{}{"a": "1"},
+			`{"a":"1"}`,
+			map[string]interface{}{"a": "1"},
+			`{"b":"2"}`,
+			map[string]interface{}{"b": "2"},
+			`{"b":"2"}`,
+			map[string]interface{}{"b": "2"},
+		},
+		{
+			"buffer req, json-parse req",
+			eval.BufferRequest | eval.JSONParseRequest,
+			"application/json",
+			`{"a":"1"}`,
+			map[string]interface{}{"a": "1"},
+			`{"a":"1"}`,
+			map[string]interface{}{"a": "1"},
+			nil,
+			nil,
+			nil,
+			nil,
+		},
+		{
+			"buffer resp, json-parse resp",
+			eval.BufferResponse | eval.JSONParseResponse,
+			"application/json",
+			nil,
+			nil,
+			nil,
+			nil,
+			`{"b":"2"}`,
+			map[string]interface{}{"b": "2"},
+			`{"b":"2"}`,
+			map[string]interface{}{"b": "2"},
+		},
+		{
+			"buffer both, don't json-parse",
+			eval.BufferRequest | eval.BufferResponse,
+			"application/json",
+			`{"a":"1"}`,
+			map[string]interface{}{},
+			`{"a":"1"}`,
+			map[string]interface{}{},
+			`{"b":"2"}`,
+			map[string]interface{}{},
+			`{"b":"2"}`,
+			map[string]interface{}{},
+		},
+		{
+			"don't buffer, don't json-parse",
+			eval.BufferNone,
+			"application/json",
+			nil,
+			nil,
+			nil,
+			nil,
+			nil,
+			nil,
+			nil,
+			nil,
+		},
+		{
+			"don't buffer, json-parse both",
+			eval.BufferNone,
+			"application/json",
+			nil,
+			nil,
+			nil,
+			nil,
+			nil,
+			nil,
+			nil,
+			nil,
+		},
+		{
+			"buffer both, json-parse both, text/plain",
+			eval.BufferRequest | eval.JSONParseRequest | eval.BufferResponse | eval.JSONParseResponse,
+			"text/plain",
+			`{"a":"1"}`,
+			map[string]interface{}{},
+			`{"a":"1"}`,
+			map[string]interface{}{},
+			`{"b":"2"}`,
+			map[string]interface{}{},
+			`{"b":"2"}`,
+			map[string]interface{}{},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(subT *testing.T) {
+			helper := test.New(subT)
+
+			req := httptest.NewRequest(http.MethodPost, "/test", io.NopCloser(strings.NewReader(`{"a":"1"}`)))
+			req.Header.Set("Content-Type", tt.contentType)
+			helper.Must(eval.SetGetBody(req, eval.BufferRequest, 512))
+
+			ctx := context.WithValue(req.Context(), request.BufferOptions, tt.bufferOptions)
+			resp := &http.Response{
+				Body: io.NopCloser(strings.NewReader(`{"b":"2"}`)),
+				Header: http.Header{
+					"Content-Type": []string{tt.contentType},
+				},
+				Request: req.WithContext(ctx),
+			}
+
+			hclContext := baseCtx.WithBeresp(resp, cty.NilVal, true).HCLContext()
+
+			beRequests := seetie.ValueToMap(hclContext.Variables[eval.BackendRequests])
+			defaultRequest := beRequests["default"].(map[string]interface{})
+			if defaultRequest["body"] != tt.expBereqsDefBody {
+				subT.Errorf("backend_requests.default.body expected: %#v, got: %#v", tt.expBereqsDefBody, defaultRequest["body"])
+			}
+			if diff := cmp.Diff(defaultRequest["json_body"], tt.expBereqsDefJsonBody); diff != "" {
+				subT.Errorf("backend_requests.default.json_body expected: %#v, got: %#v", tt.expBereqsDefJsonBody, defaultRequest["json_body"])
+			}
+
+			beRequest := seetie.ValueToMap(hclContext.Variables[eval.BackendRequest])
+			if beRequest["body"] != tt.expBereqBody {
+				subT.Errorf("backend_request.body expected: %#v, got: %#v", tt.expBereqBody, beRequest["body"])
+			}
+			if diff := cmp.Diff(beRequest["json_body"], tt.expBereqJsonBody); diff != "" {
+				subT.Errorf("backend_request.json_body expected: %#v, got: %#v", tt.expBereqJsonBody, beRequest["json_body"])
+			}
+
+			beResponses := seetie.ValueToMap(hclContext.Variables[eval.BackendResponses])
+			defaultResponse := beResponses["default"].(map[string]interface{})
+			if defaultResponse["body"] != tt.expBerespsDefBody {
+				subT.Errorf("backend_responses.default.body expected: %#v, got: %#v", tt.expBerespsDefBody, defaultResponse["body"])
+			}
+			if diff := cmp.Diff(defaultResponse["json_body"], tt.expBerespsDefJsonBody); diff != "" {
+				subT.Errorf("backend_responses.default.json_body expected: %#v, got: %#v", tt.expBerespsDefJsonBody, defaultResponse["json_body"])
+			}
+
+			beResponse := seetie.ValueToMap(hclContext.Variables[eval.BackendResponse])
+			if beResponse["body"] != tt.expBerespBody {
+				subT.Errorf("backend_response.body expected: %#v, got: %#v", tt.expBerespBody, beResponse["body"])
+			}
+			if diff := cmp.Diff(beResponse["json_body"], tt.expBerespJsonBody); diff != "" {
+				subT.Errorf("backend_response.json_body expected: %#v, got: %#v", tt.expBerespJsonBody, beResponse["json_body"])
 			}
 		})
 	}